Moved shared libs into iam-persistence
