+++ iam-login-service/src/main/java/it/infn/mw/iam/core/web/aup/EnforceAupFilter.java
@@ -34,11 +34,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import it.infn.mw.iam.api.account.AccountUtils;
-import it.infn.mw.iam.api.aup.AUPSignatureCheckService;
 import it.infn.mw.iam.api.aup.error.AupNotFoundError;
 import it.infn.mw.iam.persistence.model.IamAccount;
 import it.infn.mw.iam.persistence.model.IamAup;
 import it.infn.mw.iam.persistence.repository.IamAupRepository;
+import it.infn.mw.iam.service.aup.AUPSignatureCheckService;
 
 
 public class EnforceAupFilter implements Filter {
