+++ iam-persistence/src/main/java/it/infn/mw/iam/core/time/TimeProvider.java
@@ -13,8 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package it.infn.mw.iam.service.aup;
+package it.infn.mw.iam.core.time;
 
+/**
+ * Time provider interface.
+ *
+ */
 @FunctionalInterface
 public interface TimeProvider {
 
@@ -22,7 +26,7 @@ public interface TimeProvider {
    * Returns the current time in milliseconds
    * 
    * @return the difference, measured in milliseconds, between the current time and midnight,
-   *         January 1, 1970 UTC.
+   * January 1, 1970 UTC.
    * 
    * @see System#currentTimeMillis()
    */
